Add high performance version of stereo2mono depth buffer adjustment
This takes a new approach on using stereo2mono to improve the accuracy
of the automatic HUD adjustment from the depth buffer. Rather than using
stereo2mono on the whole depth buffer (which can be a performance killer
in SLI above 1080p), this performs this calculation individually on each
eye and writes the result into a 1x1 pixel texture. That 1x1 texture is
then sent to both eyes using stereo2mono and the hud adjustment uses the
results from both eyes to decide on the final depth of the HUD.

It sets StereoFlagsDX10 = 0x00000008 in the driver profile to allow
stereo2mono to work in SLI (turns out other games where this already
worked did so because they had a compatibility mode profile which
implies this setting).