Dreamfall Chapters: Make HUD depth follow icons
The automatic adjustment using the mouse cursor position was just for
demonstration - this is closer to the final adjustment we want for this
game, where the HUD depth will follow the position of certain icons on
the screen. We can prioritise specific icons - e.g. when making a
selection we prioritise the large selection circle so that the entire
HUD depth, including the mouse cursor will be matched with the selection
HUD position/depth.

This uses the vertex buffer of the floating icon shader to locate the
center of the appropriate icon.