Dreamfall Chapters: Separate subtitle adjustment from main HUD
We don't want the subtitles at the standard HUD depth, since that will
usually put them too deep and difficult to read. Instead we probe
several spots on the depth buffer to decide what depth to place these -
we don't use the full auto crosshair adjust routine, because that is too
precise for our needs - we just need an approximate minimum depth around
the bottom of the screen.