Dreamfall Chapters: Make auto HUD adjustment rest on top of ghosts
The ghosts (and likely other transparent objects) are not rendered to
the depth buffer, so the auto HUD adjustment ignores them when
determining the depth to use. Since the ghosts can have interaction
icons we want them taken into account, so this commit uses some brand
new functionality in 3DMigoto to change the depth/stencil state of the
game when these are drawn and renders them to a copy of the depth
buffer, with the pixel shader disabled so as not to draw them a second
time to the render target. Then, if that copy has been created it will
be used in place of the original depth buffer when performing HUD
calculations, or the original will be used if there were no ghosts on
screen.

The nice thing about this is that it requires zero shader tweaks or new
shaders and only requires some tricks in the d3dx.ini.