Dreamfall Chapters: Debugging
This commit is mostly for informative purposes. It adds a bunch of
debugging to pull and display values from the hud analysis shader live
to track down why the HUD adjustment doesn't work on the ghosts before
hitting F10.

The results show:
cb0 is fine
vb0 is ok, though W goes to zero when the icon is exactly on the ghosts
The code is executed
find_quad_center() works fine
calc_0bd32bb622c2d611_pos() is bad and returns 0

This is still not entirely clear to my why this happens, but I have a
solution...