Fix Cinematic AO shader
This shader was a little tricky to locate the correct place to insert a
fix, but I noticed that it explicitly read from the off-center part of
the projection matrix, suggesting that it was already stereo capable.

It does not read from the separation*convergence part of the matrix
though (as that would be done in the view matrix with native Unity
stereo support), so this may not be the complete fix - but showing the
output live with a debug shader looks fine.

There is still a glitch with a similar appearance to this one, but it
seems like it might be coming from some other shader.