Add inventory HUD adjustment, two modes
1. When first opened the items are displayed inside circles at the
bottom of the screen and they already have depth that pushes them quite
a way into the scene. Since the items can have all sorts of vertex
shaders that are also used for the same items outside of the inventory
(at least, they were in the old DX9 build IIRC) we don't have a good
option to override the inventory depth, so we have to set the rest of
the HUD (mouse cursor in particular) to match the existing inventory
depth, but don't adjust HUD elements that already had depth.

2. After right clicking on an item it is brought forward uncomfortably
close and enlarged. The HUD also has several elements being drawn in
this situation that have an existing depth, but that depth is far too
deep, especially compared to the now uncomfortable item. In this
situation we have to bring all the HUD elements forward to line up with
the item, regardless of whether they already had depth or not. We really
need to lower the separation value while in this mode.