Rerun lighting autofix to get assembly versions for better perfomance
Wipe out all the HLSL versions of the lighting fix then re-run the
autofix to add the new asm versions and merge in the new lines in the
d3dx.ini with vimdiff.

This improves the framerate on my system between 5 and 15fps depending
on the location! Still about 5fps off the framerate achieved with
show_original, so there might still be room for improvement.