Dreamfall Chapters: Re-run autofix to replace MVP.I with VP.I
This is largely an experiment, to see how much framerate can be saved by
removing one matrix multiply from almost every shader. Individually, or
even in large groups this would be negligible... But this is a
particularly large group, so could it add up?

And the answer: Not really - maybe 1fps, but it's in the noise.