Optimise inverse matrix buffer accesses
This is another experiment, to see how significant the performance hit
is of copying the MVP.I / VP.I buffers into a constant buffer slot every
time they are accessed. It performs the full copy into a constant buffer
once, then references the copy from every access.

The result: About 5fps. Most importantly this gets us within 1fps of the
performance observed while holding the show_original key, so this is a
win.