Completely expunge MVP.I matrix
No longer need this, so get rid of it. Oddly, there were three
references left to it in the d3dx.ini - I had tried to be careful, but
they must have slipped through the cracks. Their corresponding shaders
were updated to use VP.I, so go ahead and update them in the d3dx.ini as
well.