Disable autofix on light shafts + list all shaders
The autofix causes glitches on this shader, so replace it with a vanilla
shader. List all the related light shaft shaders in the d3dx.ini so we
can start playing around with them - But, I've spent a lot of time on
these in the past in Helix Mod without success. I have some hope that
3DMigoto has (or can be extended to have) the capabilities necessary to
fix part of these, but I'm still not certain if they are entirely
fixable, so if progress is too slow I'll put these down.