Disable light shafts (and add debugging)
The light shafts in this game are still just as tricky as ever. Having
3DMigoto helps a lot, and two debugging shaders are included to
visualise the calculated light position and display input values and
working out live, and that has allowed me to get close, but there are
still some very strange things going on...

For instance, the far clipping plane we calculate in the light shaft
local coordinates varies (between 70 - 130, so not just floating point
error)! This may just be an error in my calculations, but I can't find
anything wrong. The calculated light position seems pretty close
(certainly closer than any of my previous blind attempts with Helix
Mod), and it's position in local coordinates does seem to remain pretty
close to constant independent of the camera (which is consistent with
what I think it should do), but as the camera is rotated it can be seen
to get a little too close and move to the wrong side of the wall - it
might be good enough, but the correct pattern to fix these is still
unknown anyway.

This adds handling=skip to all the light shaft shaders to disable them.