Volumetric Light Shafts: Clear render targets to fix rectangular holes
The technique we used to replace the render targets from the game is a
bit of a hack, since the game still has pointers to the original render
targets, and when it goes to clear them it will still clear the original
targets, not our replacements (If we forced the stereo mode with
TextureOverride this would not be a problem, but doing it dynamically
like this through resource copying means we can't (yet) switch out the
game's reference).

But, we know which render targets the game tries to clear, so we can
easily clear our replacement render targets as well. There is a catch
here though, since the passes that need to clear a render target may be
called multiple times per light shaft, but we only want to clear the
render target the first time it is called per light shaft per frame -
any more and we won't render the light shaft. To solve this, we set a
flag when any of the Coords shaders are used indicating that a new light
shaft is being drawn, check the flag is set on any pass that needs a
render target to be cleared, then clear the flag after clearing the
render target. Since we still lack conditional logic in the render
targets, we just do that check in the vertex shader and unconditionally
set and clear the flags in the command lists.