The volumetric light shafts - The bane of my existence, are finally working :)
Finally cracked it =D

These light shafts have been the bane of my existence, being one of the
earliest effects I started looking at back when this game was using DX9
and quite possibly the all time hardest effect to fix, and until now
I've had to release all the DFC fixes with these disabled.

This uses several new techniques to fix these:

Firstly, none of the shaders are hand edited any more - these are all
patched with asmtool to redirect accesses from cb0 ($Gloabals) to t100
instead. t100 is a stereo copy of cb0 that is filled out using a compute
shader to adjust the 2D light position on the screen, the camera
position in local coordinates relative to the light shaft, and the
camera frustum also in local coordinates. Without any matrix to
facilitate converting to local coordinates, it uses the camera position
and frustum to determine the camera X, Y and Z axis and X, Y and Z
scaling (yes, the scaling is not uniform in every direction) in local
coordinates, and projects the light position to the far clipping plane
then finds where it intersects local X = Y = 0 to determine it's
distance from the camera in the Z axis.

Secondly, a number of the render targets used by the light shafts need
to be forced to stereo, but as the descriptions vary wildly (differing
sizes for each light shaft), and some of them cannot be distinguished
from other render targets that must remain mono (Not confirmed in DX11,
but IIRC the shadow map needs to be mono but could have the same
description as DepthEpi which needed to be stereo - one of the reasons I
don't think this effect can be fixed with Helix Mod). Rather than
matching these with [TextureOverride] we match the shaders they are used
from and switch them out for stereo versions with 3DMigoto's arbitrary
resource copying support. This has a downside since the game still holds
references to the original render targets and will not clear ours - this
will be fixed in the next patch.

This sets StereoFlagsDX10 = 0x00004008 so as to work in single GPU as
well (the 8 was already there for stereo2mono to work in SLI).