Volumetric Light Shafts: Fix incorrect clipping on distant shafts
When the camera is close to a light shaft it renders using a full screen
quad, but from further away it is drawn in the world, which the driver
did not stereoise (and the auto halo fix broke as a result). Manually
apply the stereo correction to both position and texcoord outputs to
move it into 3D and fix the clipping.