Volumetric Light Shafts: Disable render target redirection from common shader
One of the pixel shaders amongst the light shafts is just a simple
shader used to clear a render target, and is also used by two other
effects. Remove this one from the list of shaders that participate in
render target redirection so that we don't interfere with other effects
using the same shader. We already replace this with our own render
target clearing procedure anyway, and I don't think it is ever used in
DX.