Fix full screen fade to white detection (undefined W in vertex buffers)
This gotcha got me again - the W value in the vertex buffer seems to be
undefined, which can throw out the position calculations if we use it,
and that prevented us from detecting the full screen fade to white
texture in the intro cinematic.

Always set W=1 when calculating a position from a vertex buffer!