Add a minimum subtitle depth
In the conversation with the dreaming girl in the prologue, for several
seconds after making a choice some invisible item is obstructing the
lower part of the depth buffer where the subtitle adjustment is probing,
making the subtitles pop out too far.

Add a minimum subtitle adjustment to make sure that the subtitles are
never popped out more than the given distance.

Also add a third pass to the HUD debugging shader that shows the depth
at the mouse cursor - this makes it easy to see issues like this (pause
the game and/or bring up the steam overlay to unlock the cursor from the
center of the screen), and also can identify if transparent surfaces are
being drawn on the depth buffer or not.