Fix transparent depth buffer being falsely obscured popping out subtitles too far
This fixes the issues with the subtitles being popped out too far by
using a brand new feature of 3DMigoto to abort the transparent depth
buffer shader when the wrong depth buffer is in use. The correct depth
buffer that is used while rendering transparent objects has bind flags
to allow it to be used as a shader resource, while the wrong one that is
used while rendering the consequences HUD can only be used as a depth
buffer, and we take advantage of that to match them by hash and abort if
the wrong one is in use.