Add allow_duplicate_hash to machine generated sections
This is a new feature in the next version of 3DMigoto that allows
ShaderOverride sections with duplicate hashes, provided that all
duplicate sections consent to this with 'allow_duplicate_hash = true',
and have unique section names. Do a search and replace to add this tag
to all sections added with asmtool, so that we can add sections for the
transparent depth buffers without having to merge the two together.