Bring back DFC wet glass transparent shader
Hmmm, this one has Z writing enabled, so my script didn't pick it up...
Also confirmed that there is a transparent depth buffer that the game
uses as well, but it lacks some things (e.g. has the glass and
holographic sculpture in Roman's office, but lacks the screen on his
desk).