Bring back political ad shader to transparent list
This was one of the two shaders I blacklisted for messing up the
subtitles while the consequence text was displayed, but turns out it was
one I had earlier included and do need.  Add it back in and see what
happens - it might have been the other one that messed up the subtitles.

Had another look, and the hash of the depth buffer has changed even
though the description has not! May need to investigate that...