Remove autofix from fresnel distort shaders
These are used for the holographic displays in The Hand That Feeds, but
the effect they are going for is fundamentally broken in 3D and the
stereo correction makes it look weird from certain angles. Remove the
autofix from these shaders to make it look tolerable (still wrong, but
that's not my problem).