Fix Screen Space Reflections
This is a relatively easy SSR shader to fix, however while the shader is
used all over the game, the effect is subtle in most places and hard to
be sure if it is fixed 100% or just approximated. Makes a larger
difference to horizontal walls reflecting Zoe or lights, but also fixes
the reflection lifting off the water when the camera is pointing down.

The SSR reflections in The Hand That Feeds are still weird, but that is
the case in 2D as well, so not my problem