Invert transparent depth buffer test + add more transparent shaders
This adds 'Sonic Ether_Particles_Additive' to the transparent list to
fix the screen in the lab with KidBot, however doing so broke the HUD
from certain angles, caused the screen to render incorrectly and
possibly related, Kitbot's face was blank - it seems like our custom
shader was being run when it shouldn't be and messing things up.

This inverts the depth buffer test - rather than bailing out when a
known bad depth buffer is in use, run the custom shader from the known
good depth buffer instead. That prevents the redirect from happening if
there are any other bad depth buffers, but also means if no depth buffer
is bound at all we won't try the redirect.