Blacklist transparent shader to fix HUD in Abnaxis Dream
This shader is used for puffy white clouds that we don't want to take
into account for the HUD depth, so disable it.

Will need to go back and re-visit the lab to see if the screen was
affected by this one, or one of the other variants of this shader.