Add debugging option to disable desaturation in menus
This is an optional replacement of the original desaturate shader that
effectively disables it - we can't simply skip it in this case. This is
useful when the game is paused to remove the desaturation from the menu,
and in combination with disabling the HUD shaders can allow whatever is
behind to be seen.