Fix lighting in Book 3 Chapter 7 (update and re-run autofix)
This updates the assembly lighting fix to use the FOV passed from the
vertex shader if available, and only use the copied in inverse VP matrix
if not. This makes the lighting fix more reliable in case something went
wrong copying the inverse view-projection matrix, and fixes the lighting
in Book 3 Chapter 7.