Fix sun depth in book 4 chapter 8 (risky)
First time we see the real sun in this game is in book 4!

This is the flare shader that is broken in many Unity games, but usually
it is just broken by driver heuristics, in this case it is really at the
wrong depth. This adds checks to try to nail this down and avoid
breaking anywhere else the shader might have been used, but it is still
risky and I'll need to go back and check a few other places.

- Check depth is within a certain range (unfortunately it is a large
  range...)
- Check pixel shader matches flare shader
- Check texture matches