Fix lighting Book 5 scene 1 by copying UnityPerCameraRare by ref
This restores the UnityPerCameraRare that had been stripped out by
cleanup_unity_shaders in the previous version of the game, adds an extra
clause for the soft directional lighting shader this games uses, and
switches them all to copy by reference without any copy limitations.

Collectively, these changes fix the lighting in the first scene in Book
5 that must be rendering another scene before the frame throwing things
out.

The light shafts are still broken in this scene for much the same reason
:(