Use reset_per_frame_limits to fix lighting & light shafts at start of book 5
This uses yet another brand new feature in 3DMigoto to fix the lighting
& light shafts in the cutscene right at the start of book 5. This
cutscene is rendering two separate scenes off-screen before it renders
the main scene, and grabbing the constant buffers from those scenes will
mess up the fixes in the main scene.

This resets the per frame counters on each of the resources we copy and
matrix inverse custom shader every time a depth target is cleared. That
could lead to these being run a few more times than necessary, but still
far less than if we just allowed them to be copied on every draw call.

This also undoes copying UnityPerCameraRare by reference - while that
seems to still be working and is probably fine, copying UnityPerCamera
by reference messes up the light shafts, so playing it safe.