Reset CameraDepthTexture copy limit on depth buffer clear
This fixes the subtitle depth in parts of Book 5, though I am a tad
concerned about changing this this late into the game. Another option
that should perform slightly better would be to always copy the depth
buffer by reference then do a full copy of that the first time we use it
in a HUD shader, however of the two I judge that to be slightly more
risky this late into the game so I'm opting to reset the copy limit.